Moves EventCreateForm to GroupHeader from AppControl, and modifies GroupHeader and EventCreaateForm so that user create an event by clicking the button on GroupView page
