+++ client/we_meet_client/src/components/ui-components/group-view/GroupHeader.js
@@ -1,18 +1,52 @@
 import React, {Component} from "react";
-import moment from "moment";
 import { FillButton } from 'tailwind-react-ui';
 import axios from "axios";
-
+import EventForm from '../../Events/EventForm'
 
 
 class GroupHeader extends Component {
+  state = {
+    showCreateEventForm: false,
+    navBarNotification: ''
+  }
 
   async joinGroup() {
-    const credentials = { 'access-token': localStorage.getItem('access-token'), 'token-type': localStorage.getItem('token-type'), 'client': localStorage.getItem('client'), 'expiry': localStorage.getItem('expiry'), 'uid': localStorage.getItem('uid'), }
-
+    const credentials = { 
+      'access-token': localStorage.getItem('access-token'), 
+      'token-type': localStorage.getItem('token-type'), 
+      'client': localStorage.getItem('client'), 
+      'expiry': localStorage.getItem('expiry'), 
+      'uid': localStorage.getItem('uid')
+    }
     const response = await axios.post(`http://localhost:3000/groups/${this.props.group.id}/memberships`, {}, { headers: credentials }) 
-    console.log(response)
   }
+
+  showCreateEventForm = () => {
+    this.setState({ showCreateEventForm: true })
+  }
+
+  hideCreateEventForm=() => {
+    this.setState({ showCreateEventForm: false })
+  }
+
+   createEventHandler = async(event) => {
+    
+
+    const credentials = {
+     'access-token': localStorage.getItem('access-token'),
+     'token-type': localStorage.getItem('token-type'),
+     'client': localStorage.getItem('client'),
+     'expiry': localStorage.getItem('expiry'),
+     'uid': localStorage.getItem('uid'),
+    }
+    console.log(this.props.group)
+    const response = await axios.post(`http://localhost:3000/groups/${this.props.group.id}/events`, { event }, { headers: credentials })
+    this.setState({ navBarNotification: response.data.message, showCreateEventForm: false })
+  }
+
+  
+
+
   render() {
     let group = this.props.group;
 
@@ -67,12 +101,16 @@ class GroupHeader extends Component {
                   Join group
               </FillButton>
               <FillButton
+                  onClick={this.showCreateEventForm}
                   brand='secondary'
                   large
                   style={{ position:"absolute", right:"16.5rem", width:"300px", marginTop:"15rem", fontSize:"24px"}}
                 >
                   Create event
               </FillButton>
+
+              { this.state.showCreateEventForm && <EventForm hideCreateEventForm={ this.hideCreateEventForm} createEventHandler={this.createEventHandler}/> }
+
             </div>
           {/* </div> */}
         </div>
